stuck on invalid auth move to front end

+++ Keepr/Services/VaultsService.cs
@@ -14,6 +14,18 @@ namespace Keepr.Services
       _vaultrepo = vaultrepo;
     }
 
+    private Vault AllowAccess(int id, string userId)
+    {
+      Vault original = Get(id);
+      if (original.IsPrivate == true || original.CreatorId != userId)
+      {
+        throw new Exception("Access Denied");
+      }
+      return original;
+    }
+
+
+
     internal Vault Create(Vault vaultData)
     {
       return _vaultrepo.Create(vaultData);
@@ -26,27 +38,38 @@ namespace Keepr.Services
       {
         throw new Exception("Invalid Id");
       }
-      else if (found.IsPrivate == true)
-      {
-        throw new Exception("You cannot access this vault");
-      }
+
+
       return found;
     }
 
-    internal Vault Edit(Vault vaultData)
+
+
+    internal Vault Edit(Vault vaultData, string userId)
     {
-      Vault original = Get(vaultData.Id);
+      Vault original = AllowAccess(vaultData.Id, vaultData.CreatorId);
+
 
       original.Name = vaultData.Name ?? original.Name;
       original.Description = vaultData.Description ?? original.Description;
+      original.IsPrivate = vaultData.IsPrivate ?? original.IsPrivate;
 
       _vaultrepo.Edit(original);
+      original.UpdatedAt = new DateTime();
       return original;
     }
 
-    internal void Delete(int id)
+    internal void Delete(int id, string userId)
     {
       Vault foundVault = Get(id);
+      if (foundVault.CreatorId != userId)
+      {
+        throw new Exception("You cannot delete this");
+      }
+      else if (foundVault.IsPrivate == true && foundVault.CreatorId != userId)
+      {
+        throw new Exception("Access Denied");
+      }
       _vaultrepo.Delete(id);
     }
